+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - fcoe_network support for API300
   - network_set support for API300
   - datacenter support for API300
+  - fabric new resource support for API300 Synergy
 
 ### 1.1.0
   - Add support for client ENV variables